feat(extractor): optimise readability
